* origin/main:
  Adversarial
  refactor(kibana/security_detection_rule): merge duplicate CommonCreateProps/CommonUpdateProps into CommonRuleProps (#2795)
  refactor(elasticsearch): export IsNotFoundElasticsearchError and remove acctest duplicate (#2791)
  feat(elasticsearch): migrate transform to Plugin Framework entitycore envelope (#2757)
  refactor(kibana/agentbuilder): extract shared version-gating helper (#2792)
  feat(elasticsearch): migrate component_template to Plugin Framework entitycore envelope (#2749)

+++ internal/models/models.go
@@ -120,7 +120,7 @@ type ComponentTemplate struct {
 	Name     string         `json:"-"`
 	Meta     map[string]any `json:"_meta,omitempty"`
 	Template *Template      `json:"template,omitempty"`
-	Version  *int           `json:"version,omitempty"`
+	Version  *int64         `json:"version,omitempty"`
 }
 
 type ComponentTemplateResponse struct {
